+++ Firestore/core/src/firebase/firestore/remote/grpc_stream_operation.h
